Separated files using cogs, stats.py doesnt load, commands are not recognized
